requirements.txt
pandas
pymongo
flask
dnspython
progressbar2